+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/directory/FDBDirectoryManager.java
@@ -142,7 +142,7 @@ public class FDBDirectoryManager implements AutoCloseable {
             }
         } else {
             AtomicReference<LucenePartitionInfoProto.LucenePartitionInfo> lastPartitionInfo = new AtomicReference<>();
-            return AsyncUtil.whileTrue(() -> getNextPartitionInfo(groupingKey, agileContext, lastPartitionInfo)
+            return AsyncUtil.whileTrue(() -> getNextOlderPartitionInfo(groupingKey, agileContext, lastPartitionInfo)
                     .thenApply(partitionId -> mergePartition(analyzerWrapper, groupingKey, agileContext, partitionId)));
         }
     }
@@ -169,10 +169,10 @@ public class FDBDirectoryManager implements AutoCloseable {
         }
     }
 
-    private CompletableFuture<Integer> getNextPartitionInfo(final Tuple groupingKey, final AgilityContext agileContext,
-                                                            final AtomicReference<LucenePartitionInfoProto.LucenePartitionInfo> lastPartitionInfo) {
-        return agileContext.apply(context -> LucenePartitioner.getNextPartitionInfo(
-                        context, groupingKey, lastPartitionInfo.get(), state.indexSubspace)
+    private CompletableFuture<Integer> getNextOlderPartitionInfo(final Tuple groupingKey, final AgilityContext agileContext,
+                                                                 final AtomicReference<LucenePartitionInfoProto.LucenePartitionInfo> lastPartitionInfo) {
+        return agileContext.apply(context -> LucenePartitioner.getNextOlderPartitionInfo(
+                        context, groupingKey, lastPartitionInfo.get() == null ? null : LucenePartitioner.getPartitionKey(lastPartitionInfo.get()), state.indexSubspace)
                 .thenApply(partitionInfo -> {
                     lastPartitionInfo.set(partitionInfo);
                     return partitionInfo == null ? null : partitionInfo.getId();